Stop events reaching the portal block if it they are on the front face
Prevents the player from right clicking or breaking the portal block
from the front face. It prevents the right clicking via the
onBlockActivated method which is called when a block is used. This
interaction will be passed through to the tile when it is required. The
block break interactions are prevented using an event handler on the
PlayerInteractionEvent.LeftClickBlock, and denys the use of a block or
item on any block that is a portal, when they attempt to do it on the
front side.

Signed-off-by: Luke Gilfoyle <boo122333@gmail.com>