Update build.gradle
This update makes it so that the genIntellijRuns task not only adds the
idea module to the run configs, but also makes it possible to reload
classes (hotswap) when running in debug mode.

This was not possible before the change because the way it was set up
meant that gradle was building the project and the output directory was
not one that idea knows exists. So changing the build.gradle to
explicitly say which directory contains the compiler output lets idea do
its class reloading goodness.

Signed-off-by: Luke Gilfoyle <boo122333@gmail.com>