Add localization to all the things
This adds a MessageKey enum which we can add to as we want to provide
the player with information. This enum prefixes the message key provided
with "message.portalmod." as to make it easy to spot this passage of
entries in the lang file, and keep them organised.

This commit also moves some files into a reference dir, this is because
the moved classes only store reference values.

Signed-off-by: Luke Gilfoyle <boo122333@gmail.com>