Add SolidFuelGenerator WIP - requires texture and model
This adds the solid fuel generator. This is a block which burns solid
fuels to create power in the form of FE (Forge Energy). This block
currently generates 20FE/t and fuels last the same amount of time as
they would in a furnace.

This commit also adds a new system for storing NBT called NBTKey. This
class contains some NBTKeys for ENERGY and ITEMS for now, and it will be
expanded in the future.

NBTKeys work by letting you pass a compoundNBT and object to the key,
in either a write or read method, which will write the object or read
the object to/from the compound respectively.

Signed-off-by: Luke Gilfoyle <boo122333@gmail.com>