Add modules to slit cannon and implement a new enum-based config system
Added an energy module and item module to the slit cannon's tile so it
can hold an ender eye and use energy to fire it.

Added a config to decide the fe required to fire an ender eye using the
slit cannon. This is what led to the config system.

The config system works by creating an enum singleton which implements
IConfigHolder, and then adding that config holder to the configHolders
in the Config class. Then you can implement the writeServer, writeClient
and/or writeCommon methods to write the configs on the matching config
spec. If you override any of these methods, you will also need to
override the relevant hasX method and return true. The recommended way
to access the values in the config enums is to write a getter for them.
See the SolidFuelGeneratorConfig for an example of how to use this
system.

Signed-off-by: Luke Gilfoyle <boo122333@gmail.com>