Fix event handlers and add whitelist for interactions with portals front
Removes the static qualifier from the common event handlers, because
they are supposed to be members of the enum instance that is being
registered to the event bus.

Adds an item tag which contains a whitelist of items that can interact
directly with the front face of a portal block. Currently that list of
items is limited to: portal linker and portal link breaker.

Signed-off-by: Luke Gilfoyle <boo122333@gmail.com>