Adds BlockState data generator
Adds block state data generator and all the things required for it.

Removes State property from solid fuel generator, as it makes a mess of
the blockstate file. In doing so i added it to the tile entity. It may
be moved back in the future when we add more models for when then
generator is active and not active, which would mean we want the
property to appear in the blockstate as it was before it was removed.

Signed-off-by: Luke Gilfoyle <boo122333@gmail.com>